# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DeClawd API
  description: >
    REST API for DeClawd, a non-custodial, AI-powered BTC prediction-market
    trading platform on Polymarket. All authenticated endpoints use an
    httpOnly JWT cookie set during the Google OAuth flow. DeClawd never
    holds private keys or seed phrases - wallet-authorizing endpoints only
    ever verify a signature produced by the user's own wallet.
  version: '1.0'
  contact:
    name: DeClawd
servers:
  - url: http://localhost:4000
    description: Local development
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/auth/google/callback:
    get:
      tags: [Auth]
      summary: Google OAuth callback
      security: []
      description: >
        Exchanges the OAuth code for a Google identity, creates/looks up the
        User, issues access + refresh JWTs as httpOnly cookies, and redirects
        to the web app.
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to WEB_BASE_URL with session cookies set
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate access token using refresh cookie
      security: []
      responses:
        '200':
          description: New access token cookie issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEmpty'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get the current authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessBase'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserDto'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # --------------------------------------------------------------------
  # Wallet
  # --------------------------------------------------------------------
  /api/v1/wallet/nonce:
    post:
      tags: [Wallet]
      summary: Request a SIWE-style nonce/message to sign for wallet linking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address, chainId]
              properties:
                address:
                  type: string
                  example: '0x1234567890abcdef1234567890abcdef12345678'
                chainId:
                  type: integer
                  example: 137
      responses:
        '200':
          description: Message to be signed by the wallet
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessBase'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          nonce:
                            type: string
                          message:
                            type: string
                          expiresAt:
                            type: string
                            format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/wallet/link:
    post:
      tags: [Wallet]
      summary: Link a wallet after verifying its signature over the nonce message
      description: >
        Never trusts a wallet address without a verified signature. The
        server independently recovers the signer from `signature` and
        `message` and rejects the request on any mismatch.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address, chainId, provider, message, signature]
              properties:
                address:
                  type: string
                chainId:
                  type: integer
                provider:
                  type: string
                  enum: [METAMASK, COINBASE_WALLET, WALLET_CONNECT, RABBY, RAINBOW]
                message:
                  type: string
                signature:
                  type: string
                isPrimary:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Wallet linked
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessBase'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WalletDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Signature verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/wallet:
    get:
      tags: [Wallet]
      summary: List wallets linked to the current user
      responses:
        '200':
          description: Linked wallets
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccessBase'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/WalletDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/wallet/{id}:
    delete:
      tags: [Wallet]
      summary: Unlink a wallet
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Wallet unlinked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessEmpty'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # --------------------------------------------------------------------
  # Dashboard
  # --------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: declawd_access_token
      description: >
        httpOnly JWT access-token cookie set by /api/v1/auth/google/callback
        and refreshed via /api/v1/auth/refresh. Short-lived (JWT_ACCESS_TTL);
        a separate httpOnly refresh cookie (declawd_refresh_token,
        JWT_REFRESH_TTL) is used only to mint new access tokens.
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ApiSuccessBase:
      type: object
      properties:
        success:
          type: boolean
          example: true
        meta:
          type: object
          additionalProperties: true
      required: [success]
    ApiSuccessEmpty:
      allOf:
        - $ref: '#/components/schemas/ApiSuccessBase'
        - type: object
          properties:
            data:
              type: object
              nullable: true
              example: null
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - INSUFFICIENT_BALANCE
                - WALLET_VERIFICATION_FAILED
                - MARKET_NOT_ELIGIBLE
                - TRADING_ENGINE_DISABLED
                - PROVIDER_ERROR
                - INTERNAL_ERROR
            message:
              type: string
            details:
              nullable: true
          required: [code, message]
      required: [success, error]

    # ---------------- DTOs ----------------
    UserDto:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [USER, ADMIN, SUPPORT]
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, DELETED]
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
    WalletDto:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        chainId:
          type: integer
        provider:
          type: string
          enum: [METAMASK, COINBASE_WALLET, WALLET_CONNECT, RABBY, RAINBOW]
        isPrimary:
          type: boolean
        verifiedAt:
          type: string
          format: date-time
...